Ported to the 2018 lab build with CUDA 8.0.
This does not currently compile with CUDA 9.0 due to some sort of thrust incompatibility.

Note that this now depends on the pkg-config tool to return the lib and include paths. Much neater than defining CUDA_PATH in your .bashrc, but you still need to add the following on the lab build to your .bashrc to find the pkg-config results:
export PKG_CONFIG_PATH=$PKG_CONFIG_PATH:/public/devel/cuda-8.0/pkgconfig:/public/devel/cuda-9.0/pkgconfig

This might be automated in a future lab build.